+++ impeller/tessellator/BUILD.gn
@@ -15,7 +15,6 @@ impeller_component("tessellator") {
   deps = [
     "../core",
     "//flutter/fml",
-    "//third_party/libtess2",
   ]
 }
 
@@ -30,8 +29,6 @@ impeller_component("tessellator_shared") {
   sources = [
     "c/tessellator.cc",
     "c/tessellator.h",
-    "tessellator.cc",
-    "tessellator.h",
   ]
 
   deps = [
