+++ impeller/tessellator/tessellator.h
@@ -20,11 +20,6 @@ struct TESStesselator;
 
 namespace impeller {
 
-void DestroyTessellator(TESStesselator* tessellator);
-
-using CTessellator =
-    std::unique_ptr<TESStesselator, decltype(&DestroyTessellator)>;
-
 //------------------------------------------------------------------------------
 /// @brief      A utility that generates triangles of the specified fill type
 ///             given a polyline. This happens on the CPU.
@@ -71,12 +66,6 @@ class Tessellator {
   };
 
  public:
-  enum class Result {
-    kSuccess,
-    kInputError,
-    kTessellationError,
-  };
-
   /// @brief  A callback function for a |VertexGenerator| to deliver
   ///         the vertices it computes as |Point| objects.
   using TessellatedVertexProc = std::function<void(const Point& p)>;
@@ -174,32 +163,6 @@ class Tessellator {
   Tessellator();
 
   ~Tessellator();
-
-  /// @brief A callback that returns the results of the tessellation.
-  ///
-  ///        The index buffer may not be populated, in which case [indices] will
-  ///        be nullptr and indices_count will be 0.
-  using BuilderCallback = std::function<bool(const float* vertices,
-                                             size_t vertices_count,
-                                             const uint16_t* indices,
-                                             size_t indices_count)>;
-
-  //----------------------------------------------------------------------------
-  /// @brief      Generates filled triangles from the path. A callback is
-  ///             invoked once for the entire tessellation.
-  ///
-  /// @param[in]  path  The path to tessellate.
-  /// @param[in]  tolerance  The tolerance value for conversion of the path to
-  ///                        a polyline. This value is often derived from the
-  ///                        Matrix::GetMaxBasisLength of the CTM applied to the
-  ///                        path for rendering.
-  /// @param[in]  callback  The callback, return false to indicate failure.
-  ///
-  /// @return The result status of the tessellation.
-  ///
-  Tessellator::Result Tessellate(const Path& path,
-                                 Scalar tolerance,
-                                 const BuilderCallback& callback);
 
   //----------------------------------------------------------------------------
   /// @brief      Given a convex path, create a triangle fan structure.
@@ -320,9 +283,8 @@ class Tessellator {
   /// Used for polyline generation.
   std::unique_ptr<std::vector<Point>> point_buffer_;
   std::unique_ptr<std::vector<uint16_t>> index_buffer_;
-  CTessellator c_tessellator_;
 
-  // Data for variouos Circle/EllipseGenerator classes, cached per
+  // Data for various Circle/EllipseGenerator classes, cached per
   // Tessellator instance which is usually the foreground life of an app
   // if not longer.
   static constexpr size_t kCachedTrigCount = 300;
