+++ impeller/tessellator/tessellator_unittests.cc
@@ -13,73 +13,6 @@
 namespace impeller {
 namespace testing {
 
-TEST(TessellatorTest, TessellatorBuilderReturnsCorrectResultStatus) {
-  // Zero points.
-  {
-    Tessellator t;
-    auto path = PathBuilder{}.TakePath(FillType::kOdd);
-    Tessellator::Result result = t.Tessellate(
-        path, 1.0f,
-        [](const float* vertices, size_t vertices_count,
-           const uint16_t* indices, size_t indices_count) { return true; });
-
-    ASSERT_EQ(result, Tessellator::Result::kInputError);
-  }
-
-  // One point.
-  {
-    Tessellator t;
-    auto path = PathBuilder{}.LineTo({0, 0}).TakePath(FillType::kOdd);
-    Tessellator::Result result = t.Tessellate(
-        path, 1.0f,
-        [](const float* vertices, size_t vertices_count,
-           const uint16_t* indices, size_t indices_count) { return true; });
-
-    ASSERT_EQ(result, Tessellator::Result::kSuccess);
-  }
-
-  // Two points.
-  {
-    Tessellator t;
-    auto path = PathBuilder{}.AddLine({0, 0}, {0, 1}).TakePath(FillType::kOdd);
-    Tessellator::Result result = t.Tessellate(
-        path, 1.0f,
-        [](const float* vertices, size_t vertices_count,
-           const uint16_t* indices, size_t indices_count) { return true; });
-
-    ASSERT_EQ(result, Tessellator::Result::kSuccess);
-  }
-
-  // Many points.
-  {
-    Tessellator t;
-    PathBuilder builder;
-    for (int i = 0; i < 1000; i++) {
-      auto coord = i * 1.0f;
-      builder.AddLine({coord, coord}, {coord + 1, coord + 1});
-    }
-    auto path = builder.TakePath(FillType::kOdd);
-    Tessellator::Result result = t.Tessellate(
-        path, 1.0f,
-        [](const float* vertices, size_t vertices_count,
-           const uint16_t* indices, size_t indices_count) { return true; });
-
-    ASSERT_EQ(result, Tessellator::Result::kSuccess);
-  }
-
-  // Closure fails.
-  {
-    Tessellator t;
-    auto path = PathBuilder{}.AddLine({0, 0}, {0, 1}).TakePath(FillType::kOdd);
-    Tessellator::Result result = t.Tessellate(
-        path, 1.0f,
-        [](const float* vertices, size_t vertices_count,
-           const uint16_t* indices, size_t indices_count) { return false; });
-
-    ASSERT_EQ(result, Tessellator::Result::kInputError);
-  }
-}
-
 TEST(TessellatorTest, TessellateConvex) {
   {
     Tessellator t;
